+++ CHANGES.md
@@ -15,6 +15,7 @@
   and handles HTTPS URIs (#183, @gpetiot)
 - Ignore backup files when looking for README, CHANGES and LICENSE files
   (#194, @gpetiot)
+- Do not echo input characters when reading token (#199, @gpetiot)
 
 ### Removed